cp0506morning

+++ app/Livewire/Product/EditProductComponent.php
@@ -40,14 +40,16 @@ class EditProductComponent extends Component
     public $for_sell;
     public $for_rent;
     public $prices;
+    public $in_range;
+    public $price_range;
     public $address;
-    public $lat;
-    public $long;
+    // public $lat;
+    // public $long;
     public $country_id;
     public $state_id;
     public $city_id;
     public $st_id;
-    public $click_location;
+    // public $click_location;
     public $zipcode; 
 
     public $meta_keywords;
@@ -59,7 +61,7 @@ class EditProductComponent extends Component
     public $featimage;
     public $images;
     public $thumbimage;
-    Public $exchange_for;
+    public $exchange_for;
 
     public $name;
     public $slug;
@@ -77,6 +79,9 @@ class EditProductComponent extends Component
   public $newimages;
   public $newthumbimage;
   public $newfeatimage;
+  
+  public $city_id1,$city_id2,$city_id3, $price_negotiable;
+  
     public function mount($pid)
     {
         $product = Product::find($pid);
@@ -88,6 +93,8 @@ class EditProductComponent extends Component
        $this->for_sell = $product->is_sell;
        $this->for_rent = $product->is_rent;
       $this->prices =  $product->prices;
+      $this->in_range =  $product->in_range;
+      $this->price_range = $product->price_range;
      $this->address =   $product->address;
        $this->lat = $product->lat;
        $this->long  = $product->lang;
@@ -112,6 +119,12 @@ class EditProductComponent extends Component
        $this->description = $product->description ; 
        $this->s_id = $this->scategory_id;
       $this->b_id = $this->brand_id;
+      $this->st_id=$this->state_id;
+      $this->city_id3 = $product->city_id3;
+      $this->city_id2 = $product->city_id2;
+      $this->city_id1 = $product->city_id1;
+      $this->price_negotiable  = $product->price_negotiable;
+
       
 
       $Ptss= ProductAttribute::where('product_id',$pid)->get();
@@ -123,7 +136,7 @@ class EditProductComponent extends Component
 
      // dd($this->attribute_arr,$this->inputs);
      //dd($this->thumbimage);
-     //dd($this->inputs);
+    //  dd($this->inputs);
     }
 
     public function generateslug()
@@ -170,18 +183,18 @@ class EditProductComponent extends Component
         
     }
 
-    public function currentlocation()
-    {
-        // $ip = $request->ip();
-        $ip='162.159.24.227';
-        // $data = \Location::get($ip);    
-        $currentUserInfo = Location::get($ip);
-        $this->lat = $currentUserInfo->latitude ;
-        $this->long = $currentUserInfo->longitude; 
-        $this->zipcode = $currentUserInfo->zipCode;
-        //dd($currentUserInfo);
-        return;
-    }
+    // public function currentlocation(Request $request)
+    // {
+    //      $ip = $request->ip();
+    //   // $ip='162.159.24.227';
+    //     // $data = \Location::get($ip);    
+    //     $currentUserInfo = Location::get($ip);
+    //     $this->lat = $currentUserInfo->latitude ;
+    //     $this->long = $currentUserInfo->longitude; 
+    //     $this->zipcode = $currentUserInfo->zipCode;
+    //     //dd($currentUserInfo);
+    //     return;
+    // }
 
 
     public function updated($fields)
@@ -189,15 +202,16 @@ class EditProductComponent extends Component
         $this->validateOnly($fields,[
             'category_id'=>'required',
             'scategory_id'=>'required',
-            'brand_id'=>'required',
-            'modelnumber_id'=>'required',
+           // 'brand_id'=>'required',
+            //'modelnumber_id'=>'required',
             'for_exchange'=>'required',
             'for_sell'=>'required',
             'for_rent'=>'required',
             'prices'=>'required',
+            'in_range'=>'required',
             'address'=>'required',
-            'lat'=>'required',
-            'long'=>'required',
+            // 'lat'=>'required',
+            // 'long'=>'required',
             'country_id'=>'required',
             'state_id'=>'required',
             'city_id'=>'required',
@@ -243,8 +257,13 @@ class EditProductComponent extends Component
            
     //     }
     }
+    public function ranges()
+    {
+        $this->in_range=$this->in_range;
+    }
     public function updateProduct()
     {
+        //dd($this->inputs, $this->attribute_arr);
         // $this->validate([
         //     'inputs'=>'required',
         //     'attribute_arr'=>'required']);
@@ -266,16 +285,17 @@ class EditProductComponent extends Component
         $this->validate([
             'category_id'=>'required',
             'scategory_id'=>'required',
-            'brand_id'=>'required',
+           // 'brand_id'=>'required',
             
-            'modelnumber_id'=>'required',
+        //    'modelnumber_id'=>'required',
             'for_exchange'=>'required',
             'for_sell'=>'required',
             'for_rent'=>'required',
             'prices'=>'required',
+            'in_range'=>'required',
             'address'=>'required',
-            'lat'=>'required',
-            'long'=>'required',
+            // 'lat'=>'required',
+            // 'long'=>'required',
             'country_id'=>'required',
             'state_id'=>'required',
             'city_id'=>'required',
@@ -294,20 +314,30 @@ class EditProductComponent extends Component
 
             'name'=>'required',
             
-            'inputs'=>'required',
-           'attribute_arr'=>'required',
+           // 'inputs'=>'required',
+           //'attribute_arr'=>'required',
            'slug'=>'required|unique:products,slug,'.$this->pid
-        ]);
+        ],[
+            'category_id.required'=>'The category field is required.',
+            'scategory_id.required'=>'The sub-category field is required.',
+            'attribute_id.required'=>'The attribute field is required.',
+           'country_id.required'=>'The country field is required.',
+           'state_id.required'=>'The state field is required.',
+           'city_id.required'=>'The city is required.',
+           'thumbimage.required'=>'The thumnail image field is required.',
+           'featimage.required'=>'The featured image field is required.',
+           'meta_keywords.required'=>'The meta tag field is required.'
+            ]);
 
         if($this->newthumbimage)
         {
-            $this->validateOnly($fields,[
+            $this->validate([
                 'newthumbimage'=>'required|mimes:jpeg,jpg,png',
             ]);
         }
         if($this->newfeatimage)
         {
-            $this->validateOnly($fields,[
+            $this->validate([
                 'newfeatimage'=>'required|mimes:jpeg,jpg,png',
             ]);
         }
@@ -315,15 +345,24 @@ class EditProductComponent extends Component
         $product = Product::find($this->pid);
         $product->category_id= $this->category_id;
         $product->subcategory_id= $this->scategory_id;
-        $product->brand_id= $this->brand_id;
-        $product->model_id= $this->modelnumber_id;
+        if($this->brand_id){
+            $product->brand_id= $this->brand_id;
+            $product->model_id= $this->modelnumber_id;
+        }
         $product->is_exchange= $this->for_exchange;
         $product->is_sell= $this->for_sell;
         $product->is_rent= $this->for_rent;
         $product->prices= $this->prices;
+        $product->in_range=$this->in_range;
+        if($this->in_range){
+            $product->price_range=$this->price_range;
+        }
+        else{
+            $product->price_range=null;
+        }
         $product->address= $this->address;
-        $product->lat= $this->lat;
-        $product->lang= $this->long;
+        // $product->lat= $this->lat;
+        // $product->lang= $this->long;
         $product->country_id= $this->country_id;
         $product->state_id= $this->state_id;
         $product->city_id= $this->city_id;
@@ -336,17 +375,21 @@ class EditProductComponent extends Component
         $product->email_id= $this->email_id;
         if($this->newfeatimage)
         {
+            if($product->featimage){
             unlink('admin/product/feat'.'/'.$product->featimage);
+            }
             $imageNamef= Carbon::now()->timestamp.'.'.$this->newfeatimage->extension();
-            $this->newfeatimage->storeAs('product/feat',$imageName);
+            $this->newfeatimage->storeAs('product/feat',$imageNamef);
             $product->featimage = $imageNamef;
         }
 
         if($this->newthumbimage)
         {
-            unlink('admin/product/feat'.'/'.$product->thumbimage);
-            $imageNamet= Carbon::now()->timestamp.'.'.$this->thumbimage->extension();
-            $this->thumbimage->storeAs('product/thumb',$imageNamet);
+            if($product->thumbimage){
+            unlink('admin/product/thumb'.'/'.$product->thumbimage);
+            }
+            $imageNamet= Carbon::now()->timestamp.'.'.$this->newthumbimage->extension();
+            $this->newthumbimage->storeAs('product/thumb',$imageNamet);
             $product->thumbimage = $imageNamet;
         }
         if($this->newimages)
@@ -377,21 +420,29 @@ class EditProductComponent extends Component
         $product->slug=$this->slug;   
         $product->short_description =  $this->short_description;
         $product->description = $this->description; 
+        $product->price_negotiable = $this->price_negotiable;
+        $product->city_id1 = $this->city_id1;
+        $product->city_id2 = $this->city_id2;
+        $product->city_id3 = $this->city_id3;
         $product->save();    
         
-        $at = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->get();
+        $at = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->where('status','!=',3)->get();
+        //  dd($this->inputs,$at);
         foreach($at as $key => $atsd){
+            // dd($this->inputs[$atsd->id]);
+            // dd($this->inputs,$atsd->id);
             $PAT = ProductAttribute::where('product_id',$this->pid)->where('attribute_id',$atsd->id)->first();
             if(isset($PAT)){
-// dd($this->inputs,$atsd->id);
+                // dd($this->inputs,$atsd->id);
 
                 $PAT->attoption_id  = $this->inputs[$atsd->id];
                 $PAT->save();
             }else{
-            $PrAt= new ProductAttribute();
-           $PrAt->product_id = $product->id;
-           $PrAt->attribute_id = $atsd->id;
-           $PrAt->save();
+                $PrAt= new ProductAttribute();
+                $PrAt->product_id = $product->id;
+                $PrAt->attribute_id = $atsd->id;
+                $PrAt->attoption_id  = (int)$this->inputs[$atsd->id];
+                $PrAt->save();
            
             }
             
@@ -407,21 +458,22 @@ class EditProductComponent extends Component
   public function render()
     {
 
-        $categories=Category::all();
-        $scategories = Subcategory::where('category_id',$this->category_id)->get();
-        $brands = Brand ::where('category_id',$this->category_id)->where('subcategory_id',$this->scategory_id)->get();
-        $modelnumbers = ModelNumber::where('brand_id',$this->b_id)->where('category_id',$this->category_id)->where('subcategory_id',$this->s_id)->get();
-        $attributes = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->get();
-        $attributeoptions = AttributeOption::where('attribute_id',$this->attribute_id)->where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->get();
+        $categories=Category::where('status','!=',3)->get();
+        $scategories = Subcategory::where('category_id',$this->category_id)->where('status','!=',3)->get();
+        $brands = Brand ::where('category_id',$this->category_id)->where('subcategory_id',$this->scategory_id)->where('status','!=',3)->get();
+        $modelnumbers = ModelNumber::where('brand_id',$this->b_id)->where('category_id',$this->category_id)->where('subcategory_id',$this->s_id)->where('status','!=',3)->get();
+        $attributes = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->where('status','!=',3)->get();
+        $attributeoptions = AttributeOption::where('attribute_id',$this->attribute_id)->where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->where('status','!=',3)->get();
       
         $countries = Country::all();
         $states = State::where('country_id',$this->country_id)->get();
         $cities = City::where('state_id',$this->st_id)->get();
+        $citiys = City::leftJoin('states','states.id','=','cities.state_id')->where('states.country_id','101')->select('cities.*')->get();
     
         return view('livewire.admin1.product.edit-product-component',[
           'categories'=>$categories,'scategories'=>$scategories,'brands'=>$brands,
           'modelnumbers'=>$modelnumbers,'attributes'=>$attributes,'attributeoptions'=>$attributeoptions,
-          'countries'=>$countries,'states'=>$states,'cities'=>$cities])->layout('layouts.admin1');
+          'countries'=>$countries,'states'=>$states,'cities'=>$cities,'citiys'=>$citiys])->layout('layouts.admin1');
 
       //  return view('livewire.admin1.product.edit-product-component')->layout('layouts.admin1');
