+++ app/Livewire/User/EditUserProductComponent.php
@@ -40,13 +40,13 @@ class EditUserProductComponent extends Component
     public $for_rent;
     public $prices;
     public $address;
-    public $lat;
-    public $long;
+    // public $lat;
+    // public $long;
     public $country_id;
     public $state_id;
     public $city_id;
     public $st_id;
-    public $click_location;
+    // public $click_location;
     public $zipcode; 
 
     public $meta_keywords;
@@ -70,12 +70,16 @@ class EditUserProductComponent extends Component
     public $attributeoptionid;
     public $dfh;
 
+    public $in_range;
+    public $price_range;
+
     public $inputs = [];
     public $attribute_arr = [];
     public $pid;
     public $newimages;
     public $newthumbimage;
     public $newfeatimage;
+     public $city_id1,$city_id2,$city_id3, $price_negotiable;
     
     public function mount($pid)
     {
@@ -88,9 +92,11 @@ class EditUserProductComponent extends Component
         $this->for_sell = $product->is_sell;
         $this->for_rent = $product->is_rent;
         $this->prices =  $product->prices;
+        $this->in_range =  $product->in_range;
+        $this->price_range = $product->price_range;
         $this->address =   $product->address;
-        $this->lat = $product->lat;
-        $this->long  = $product->lang;
+        // $this->lat = $product->lat;
+        // $this->long  = $product->lang;
         $this->country_id = $product->country_id;
         $this->state_id = $product->state_id;
         $this->city_id = $product->city_id;
@@ -112,7 +118,12 @@ class EditUserProductComponent extends Component
         $this->description = $product->description ; 
         $this->s_id = $this->scategory_id;
         $this->b_id = $this->brand_id;
-      
+        $this->st_id=$this->state_id;
+        $this->city_id3 = $product->city_id3;
+        $this->city_id2 = $product->city_id2;
+        $this->city_id1 = $product->city_id1;
+        $this->price_negotiable  = $product->price_negotiable;
+
 
         $Ptss= ProductAttribute::where('product_id',$pid)->get();
         foreach($Ptss as $key => $Pts)
@@ -143,7 +154,7 @@ class EditUserProductComponent extends Component
         $this->option_details='';
         $this->brand_id = 0;
         $this->inputs=[];
-        $at = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->get();
+        $at = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->where('status',1)->get();
         foreach($at as $sttt)
         {
             $this->inputs[$sttt->id]=0;
@@ -170,34 +181,39 @@ class EditUserProductComponent extends Component
         
     }
 
-    public function currentlocation()
-    {
-        // $ip = $request->ip();
-        $ip='162.159.24.227';
-        // $data = \Location::get($ip);    
-        $currentUserInfo = Location::get($ip);
-        $this->lat = $currentUserInfo->latitude ;
-        $this->long = $currentUserInfo->longitude; 
-        $this->zipcode = $currentUserInfo->zipCode;
-        //dd($currentUserInfo);
-        return;
-    }
+    // public function currentlocation(Request $request)
+    // {
+    //      $ip = $request->ip();
+    //     //$ip='162.159.24.227';
+    //     // $data = \Location::get($ip);    
+    //     $currentUserInfo = Location::get($ip);
+    //     $this->lat = $currentUserInfo->latitude ;
+    //     $this->long = $currentUserInfo->longitude; 
+    //     $this->zipcode = $currentUserInfo->zipCode;
+    //     //dd($currentUserInfo);
+    //     return;
+    // }
 
+    public function ranges()
+    {
+        $this->in_range=$this->in_range;
+    }
 
     public function updated($fields)
     {
         $this->validateOnly($fields,[
             'category_id'=>'required',
             'scategory_id'=>'required',
-            'brand_id'=>'required',
-            'modelnumber_id'=>'required',
+          //  'brand_id'=>'required',
+        //    'modelnumber_id'=>'required',
             'for_exchange'=>'required',
             'for_sell'=>'required',
             'for_rent'=>'required',
+            'in_range'=>'required',
             'prices'=>'required',
             'address'=>'required',
-            'lat'=>'required',
-            'long'=>'required',
+            // 'lat'=>'required',
+            // 'long'=>'required',
             'country_id'=>'required',
             'state_id'=>'required',
             'city_id'=>'required',
@@ -220,13 +236,13 @@ class EditUserProductComponent extends Component
 
         if($this->newthumbimage)
         {
-            $this->validateOnly($fields,[
+            $this->validate([
                 'newthumbimage'=>'required|mimes:jpeg,jpg,png',
             ]);
         }
         if($this->newfeatimage)
         {
-            $this->validateOnly($fields,[
+            $this->validate([
                 'newfeatimage'=>'required|mimes:jpeg,jpg,png',
             ]);
         }
@@ -241,16 +257,17 @@ class EditUserProductComponent extends Component
         $this->validate([
             'category_id'=>'required',
             'scategory_id'=>'required',
-            'brand_id'=>'required',
+          //  'brand_id'=>'required',
             
-            'modelnumber_id'=>'required',
+        //    'modelnumber_id'=>'required',
             'for_exchange'=>'required',
             'for_sell'=>'required',
             'for_rent'=>'required',
             'prices'=>'required',
+            'in_range'=>'required',
             'address'=>'required',
-            'lat'=>'required',
-            'long'=>'required',
+            // 'lat'=>'required',
+            // 'long'=>'required',
             'country_id'=>'required',
             'state_id'=>'required',
             'city_id'=>'required',
@@ -271,17 +288,27 @@ class EditUserProductComponent extends Component
             'inputs'=>'required',
             'attribute_arr'=>'required',
             'slug'=>'required|unique:products,slug,'.$this->pid
-        ]);
+        ],[
+            'category_id.required'=>'The category field is required.',
+            'scategory_id.required'=>'The sub-category field is required.',
+            'attribute_id.required'=>'The attribute field is required.',
+           'country_id.required'=>'The country field is required.',
+           'state_id.required'=>'The state field is required.',
+           'city_id.required'=>'The city is required.',
+           'thumbimage.required'=>'The thumnail image field is required.',
+           'featimage.required'=>'The featured image field is required.',
+           'meta_keywords.required'=>'The meta tag field is required.'
+            ]);
 
         if($this->newthumbimage)
         {
-            $this->validateOnly($fields,[
+            $this->validateOnly([
                 'newthumbimage'=>'required|mimes:jpeg,jpg,png',
             ]);
         }
         if($this->newfeatimage)
         {
-            $this->validateOnly($fields,[
+            $this->validateOnly([
                 'newfeatimage'=>'required|mimes:jpeg,jpg,png',
             ]);
         }
@@ -289,15 +316,24 @@ class EditUserProductComponent extends Component
         $product = Product::find($this->pid);
         $product->category_id= $this->category_id;
         $product->subcategory_id= $this->scategory_id;
+        if($this->brand_id){
         $product->brand_id= $this->brand_id;
         $product->model_id= $this->modelnumber_id;
+        }
         $product->is_exchange= $this->for_exchange;
         $product->is_sell= $this->for_sell;
         $product->is_rent= $this->for_rent;
         $product->prices= $this->prices;
+        $product->in_range=$this->in_range;
+        if($this->in_range){
+            $product->price_range=$this->price_range;
+        }
+        else{
+            $product->price_range=null;
+        }
         $product->address= $this->address;
-        $product->lat= $this->lat;
-        $product->lang= $this->long;
+        // $product->lat= $this->lat;
+        // $product->lang= $this->long;
         $product->country_id= $this->country_id;
         $product->state_id= $this->state_id;
         $product->city_id= $this->city_id;
@@ -310,7 +346,9 @@ class EditUserProductComponent extends Component
         $product->email_id= $this->email_id;
         if($this->newfeatimage)
         {
+            if($product->featimage){
             unlink('admin/product/feat'.'/'.$product->featimage);
+            }
             $imageNamef= Carbon::now()->timestamp.'.'.$this->newfeatimage->extension();
             $this->newfeatimage->storeAs('product/feat',$imageName);
             $product->featimage = $imageNamef;
@@ -318,9 +356,11 @@ class EditUserProductComponent extends Component
 
         if($this->newthumbimage)
         {
-            unlink('admin/product/feat'.'/'.$product->thumbimage);
-            $imageNamet= Carbon::now()->timestamp.'.'.$this->thumbimage->extension();
-            $this->thumbimage->storeAs('product/thumb',$imageNamet);
+            if(thumbimage){
+            unlink('admin/product/thumb'.'/'.$product->thumbimage);
+            }
+            $imageNamet= Carbon::now()->timestamp.'.'.$this->newthumbimage->extension();
+            $this->newthumbimage->storeAs('product/thumb',$imageNamet);
             $product->thumbimage = $imageNamet;
         }
         if($this->newimages)
@@ -332,7 +372,7 @@ class EditUserProductComponent extends Component
                 {
                     if($image)
                     {
-                        unlink('admin/products/image'.'/'.$image);
+                        unlink('admin/product/image'.'/'.$image);
                     }
                 }
             }
@@ -351,9 +391,13 @@ class EditUserProductComponent extends Component
         $product->slug=$this->slug;   
         $product->short_description =  $this->short_description;
         $product->description = $this->description; 
+        $product->price_negotiable = $this->price_negotiable;
+        $product->city_id1 = $this->city_id1;
+        $product->city_id2 = $this->city_id2;
+        $product->city_id3 = $this->city_id3;
         $product->save();    
         
-        $at = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->get();
+        $at = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->where('status',1)->get();
         foreach($at as $key => $atsd){
             $PAT = ProductAttribute::where('product_id',$this->pid)->where('attribute_id',$atsd->id)->first();
             if(isset($PAT)){
@@ -363,6 +407,7 @@ class EditUserProductComponent extends Component
                 $PrAt= new ProductAttribute();
                 $PrAt->product_id = $product->id;
                 $PrAt->attribute_id = $atsd->id;
+                $PrAt->attoption_id  = (int)$this->inputs[$atsd->id];
                 $PrAt->save();
             }
         }
@@ -374,20 +419,21 @@ class EditUserProductComponent extends Component
     public function render()
     {
 
-        $categories=Category::all();
-        $scategories = Subcategory::where('category_id',$this->category_id)->get();
-        $brands = Brand ::where('category_id',$this->category_id)->where('subcategory_id',$this->scategory_id)->get();
-        $modelnumbers = ModelNumber::where('brand_id',$this->b_id)->where('category_id',$this->category_id)->where('subcategory_id',$this->s_id)->get();
-        $attributes = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->get();
-        $attributeoptions = AttributeOption::where('attribute_id',$this->attribute_id)->where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->get();
+        $categories=Category::where('status',1)->get();
+        $scategories = Subcategory::where('category_id',$this->category_id)->where('status',1)->get();
+        $brands = Brand ::where('category_id',$this->category_id)->where('subcategory_id',$this->scategory_id)->where('status',1)->get();
+        $modelnumbers = ModelNumber::where('brand_id',$this->b_id)->where('category_id',$this->category_id)->where('subcategory_id',$this->s_id)->where('status',1)->get();
+        $attributes = Attribute::where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->where('status',1)->get();
+        $attributeoptions = AttributeOption::where('attribute_id',$this->attribute_id)->where('category_id', $this->category_id)->where('subcategory_id', $this->s_id)->where('status',1)->get();
       
         $countries = Country::all();
         $states = State::where('country_id',$this->country_id)->get();
         $cities = City::where('state_id',$this->st_id)->get();
-
+        $citiys = City::leftJoin('states','states.id','=','cities.state_id')->where('states.country_id','101')->select('cities.*')->get();
+        // dd($this->st_id);
         return view('livewire.user.edit-product-component',[
           'categories'=>$categories,'scategories'=>$scategories,'brands'=>$brands,
           'modelnumbers'=>$modelnumbers,'attributes'=>$attributes,'attributeoptions'=>$attributeoptions,
-          'countries'=>$countries,'states'=>$states,'cities'=>$cities])->layout('layouts.base');
+          'countries'=>$countries,'states'=>$states,'cities'=>$cities,'citiys'=>$citiys])->layout('layouts.base');
     }
 }
